+++ delphin_6_automation/simulation.py
@@ -13,6 +13,7 @@ sys.path.insert(0, source_folder)
 
 # RiBuild Modules
 from delphin_6_automation.logging.ribuild_logger import notifiers_logger
+from delphin_6_automation.logging.ribuild_logger import ribuild_logger
 from delphin_6_automation.database_interactions import mongo_setup
 from delphin_6_automation.database_interactions.auth import auth_dict
 from delphin_6_automation.backend.simulation_worker import main
@@ -25,10 +26,12 @@ if __name__ == "__main__":
     # Setup connection
     mongo_setup.global_init(auth_dict)
     notifier_logger = notifiers_logger(__name__)
+    local_logger = ribuild_logger(__name__)
 
     try:
         main()
 
-    except Exception:
+    except Exception as err:
         print('Exited with error!')
-        notifier_logger.error('Error in main')
+        notifier_logger.exception(f'Error in main')
+        local_logger.exception('Error in main')
